crawl/requirements.txt
httpx
pypdf 
python-dotenv

chromadb
langchain-community
langchain-google-genai
langchain-chroma
google-cloud-discoveryengine
